Extract the the multiline config to allow sharing among the sources
